fix(ffe-datepicker-react): need to pad day and month in onChange
