+++ packages/ffe-datepicker-react/src/datepicker/padZero.ts
@@ -1,0 +1,6 @@
+export const padZero = (value: number) => {
+    if (value < 10) {
+        return `0${value}`;
+    }
+    return value;
+};
